% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digimat.R
\docType{methods}
\name{digimat-methods}
\alias{digimat-methods}
\alias{[,digimat-method}
\alias{dim,digimat-method}
\alias{length,digimat-method}
\title{Methods for the digimat class}
\usage{
\S4method{[}{digimat}(x, i, j, ..., drop = TRUE)

\S4method{dim}{digimat}(x)

\S4method{length}{digimat}(x)
}
\arguments{
\item{x}{A \link[=digimat-class]{digimat} object.}

\item{i}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled) and also a matrix of indices
(but only positive ones).}

\item{j}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled).}

\item{...}{Not used. Just to make \link{nargs} works.}

\item{drop}{Whether to delete the dimensions of a matrix which have
one dimension equals to 1.}
}
\description{
Methods for the digimat class

Accessor methods for class \code{digimat}. You can use positive and negative indices,
logical indices (that are recycled) and also a matrix of indices (but only
positive ones).

Dimension and type methods for class \code{digimat}.
}
