% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-dim.R
\name{dim_file}
\alias{dim_file}
\title{File dimensions}
\usage{
dim_file(file)
}
\arguments{
\item{file}{Path to file.}
}
\value{
The number of lines and columns of file
(and extra 'return' characters).
}
\description{
Number of lines and columns of file (and extra 'return' characters).
}
\examples{
tmpfile <- tempfile()
write(0:9, tmpfile, ncolumns = 2)
dim_file(tmpfile)
}
