% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmapchar.R
\docType{class}
\name{mmapchar-class}
\alias{mmapchar-class}
\alias{mmapchar_RC}
\alias{mmapchar}
\title{Class mmapchar}
\usage{
mmapchar(file, code)
}
\arguments{
\item{file}{Path of the file.}

\item{code}{Integer vector of size 256 to access integers instead of
\code{rawToChar(as.raw(0:255), multiple = TRUE)}.
See \code{mmapcharr:::CODE_012} and \code{mmapcharr:::CODE_DIGITS}.}
}
\description{
A reference class for storing and accessing matrix-like data stored on disk
in files containing only characters (digits) separated by a character.
}

\examples{
test_file <- system.file("testdata/test-windows.txt", package = "mmapcharr")
test <- mmapchar(test_file, code = mmapcharr:::CODE_012)
test[, 1:3]
test[]
readLines(test_file)

}
