% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmapchar.R
\docType{methods}
\name{mmapchar-methods}
\alias{mmapchar-methods}
\alias{[,mmapchar-method}
\alias{dim,mmapchar-method}
\alias{length,mmapchar-method}
\title{Methods for the mmapchar class}
\usage{
\S4method{[}{mmapchar}(x, i, j, ..., drop = TRUE)

\S4method{dim}{mmapchar}(x)

\S4method{length}{mmapchar}(x)
}
\arguments{
\item{x}{A \link[=mmapchar-class]{mmapchar} object.}

\item{i}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled) and also a matrix of indices
(but only positive ones).}

\item{j}{A vector of indices (or nothing). You can use positive and negative
indices, logical indices (that are recycled).}

\item{...}{Not used. Just to make \link{nargs} works.}

\item{drop}{Whether to delete the dimensions of a matrix which have
one dimension equals to 1.}
}
\description{
Methods for the mmapchar class

Accessor methods for class \code{mmapchar}. You can use positive and negative indices,
logical indices (that are recycled) and also a matrix of indices (but only
positive ones).

Dimension and type methods for class \code{mmapchar}.
}
