% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-dim.R
\name{nline}
\alias{nline}
\title{Number of lines}
\usage{
nline(file)
}
\arguments{
\item{file}{Path to file.}
}
\value{
The number of lines of the file.
}
\description{
Number of lines of a file.
}
\examples{
tmpfile <- tempfile()
write(1:5, tmpfile, ncolumns = 1)
nline(tmpfile)
}
